docs: refactor "golf" example to be exemplary for v1
- move demo code to it's own file
